react-app: move portfolio query back to screen & fix bugs
- fixed bugs where other's portfolio not viewable when logged out

refs oursky/likedao#169
