+++ react-app/src/components/PortfolioScreen/PortfolioPanel.tsx
@@ -1,8 +1,8 @@
-import React, { useCallback, useEffect } from "react";
+import React, { useCallback } from "react";
 import cn from "classnames";
 import { toast } from "react-toastify";
 import { BigNumber } from "bignumber.js";
-import { useNavigate, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import Paper from "../common/Paper/Paper";
 import LocalizedText from "../common/Localized/LocalizedText";
 import { Icon, IconType } from "../common/Icons/Icons";
@@ -13,15 +13,7 @@ import CoinBalanceCard from "../common/CoinBalanceCard/CoinBalanceCard";
 import { convertBigNumberToMillifiedIntegerString } from "../../utils/number";
 import Config from "../../config/Config";
 import { MessageID } from "../../i18n/LocaleModel";
-import {
-  isRequestStateError,
-  isRequestStateLoaded,
-} from "../../models/RequestState";
-import LoadingSpinner from "../common/LoadingSpinner/LoadingSpinner";
-import AppRoutes from "../../navigation/AppRoutes";
-import { ConnectionStatus, useWallet } from "../../providers/WalletProvider";
 import { Portfolio } from "./PortfolioScreenModel";
-import { usePortfolioQuery } from "./PortfolioScreenAPI";
 
 const ProfilePicture: React.FC<{
   profile: Portfolio["profile"];
@@ -92,61 +84,13 @@ const CoinsAmountField: React.FC<{
   );
 };
 
-const PortfolioPanel: React.FC = () => {
+const PortfolioPanel: React.FC<{ portfolio: Portfolio }> = ({ portfolio }) => {
   const { translate } = useLocale();
   const { address } = useParams();
-  const requestState = usePortfolioQuery(address);
-  const navigate = useNavigate();
-  const wallet = useWallet();
 
   const onAddressCopied = useCallback(() => {
     toast.success(translate("UserInfoPanel.addressCopied"));
   }, [translate]);
-
-  useEffect(() => {
-    if (wallet.status === ConnectionStatus.Idle) {
-      // open connect wallet dialog
-      wallet.openConnectWalletModal();
-    }
-  }, [wallet]);
-
-  useEffect(() => {
-    if (wallet.status !== ConnectionStatus.Connected) {
-      return;
-    }
-
-    if (isRequestStateError(requestState)) {
-      if (requestState.error.message === "Invalid address.") {
-        navigate(AppRoutes.ErrorInvalidAddress);
-      } else if (requestState.error.message.includes("Wallet not connected")) {
-        // do nothing
-      } else {
-        toast.error(translate("PortfolioScreen.requestState.error"));
-      }
-    } else if (isRequestStateLoaded(requestState) && !requestState.data) {
-      toast.error(translate("PortfolioScreen.requestState.noData"));
-    }
-  }, [navigate, requestState, translate, wallet]);
-
-  if (!isRequestStateLoaded(requestState)) {
-    return (
-      <Paper
-        className={cn(
-          "flex",
-          "justify-center",
-          "items-center",
-          "min-h-[486px]",
-          "sm:min-h-[330px]"
-        )}
-      >
-        <LoadingSpinner />
-      </Paper>
-    );
-  }
-
-  if (!requestState.data) return null;
-
-  const portfolio = requestState.data;
 
   return (
     <Paper className={cn("py-6", "px-5")}>
